+++ src/main/java/com/github/freedownloadhere/killauravideo/ui/interfaces/layout/IPadded.kt
@@ -1,5 +1,5 @@
 package com.github.freedownloadhere.killauravideo.ui.interfaces.layout
 
 interface IPadded {
-    var padding: Double
+    var padding: Float
 }
